skill-flow/data/skills-refined-skillclaw-36k/skillsmp/book/cli/windows/pyproject.toml
[project]
name = "book-cli"
version = "0.3.0"
description = "CLI tool for managing LaTeX book projects"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
book = "book_cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["book_cli.py", "compile_latex.py", "init_book.py", "init_latex.py", "image_gen.py"]
